+++ tests/testthat/test-partial-eval.R
@@ -41,3 +41,18 @@ test_that("fails with multi-classes", {
   x <- structure(list(), class = c('a', 'b'))
   expect_error(partial_eval(x), "Unknown input type", fixed = TRUE)
 })
+
+# across() ----------------------------------------------------------------
+
+test_that("across() translated to individual components", {
+  # test partial_eval_across() indirectly via SQL generation
+  lf <- lazy_frame(a = 1, b = 2)
+  expect_snapshot(lf %>% summarise(across(everything(), "log")))
+  expect_snapshot(lf %>% summarise(across(everything(), log)))
+  expect_snapshot(lf %>% summarise(across(everything(), list(log))))
+
+  expect_snapshot(lf %>% summarise(across(everything(), "log", base = 2)))
+
+  expect_snapshot(lf %>% summarise(across(everything(), c("log", "exp"))))
+  expect_snapshot(lf %>% summarise(across(everything(), c("log", "exp"), .names = "{.fn}_{.col}")))
+})